only drop the versions collection if live doesn't contain the collection

+++ deployment/staging/push/before/copy_live_db.rb
@@ -36,9 +36,17 @@ MongoTools.dump(
 
 target_db = Db.from_uri(target_db_uri)
 
-puts "deleting migrations - because we want them all to run"
-`mongo #{target_db.host}:#{target_db.port}/#{target_db.name} -u #{target_db.username} -p #{target_db.password} --eval "db.mongo_migrator_versions.drop();"`
-raise "error dropping versions" unless $?.to_i == 0
+## Get the versions count
+count_raw = `mongo #{live_db.host}:#{live_db.port}/#{live_db.name} -u #{live_db.username} -p #{live_db.password} --eval "db.mongo_migrator_versions.count();"`
+raise "error getting versions count" unless $?.to_i == 0
+count = count_raw.chomp[-1]
+
+if count == "0"
+  puts "deleting migrations - because they aren't going to get overrwritten"
+  `mongo #{target_db.host}:#{target_db.port}/#{target_db.name} -u #{target_db.username} -p #{target_db.password} --eval "db.mongo_migrator_versions.drop();"`
+  raise "error dropping versions" unless $?.to_i == 0
+end
+
